Revert "More Shader Optimisations: (Mix Integer with Float Operations)"
This reverts commit 829ff599d9586fa5a75ac4f287e5652162ec085e.

+++ src/main/java/net/vulkanmod/vulkan/VRenderSystem.java
@@ -7,7 +7,6 @@ import net.minecraft.client.render.*;
 import net.minecraft.client.util.Window;
 import net.minecraft.util.math.Matrix4f;
 import org.joml.Vector2f;
-import org.joml.Vector2i;
 import org.lwjgl.system.MemoryUtil;
 import oshi.SystemInfo;
 import oshi.hardware.CentralProcessor;
@@ -257,9 +256,9 @@ public class VRenderSystem {
         Drawer.setDepthBias(0.0F, 0.0F);
     }
 
-    public static Vector2i getScreenSize() {
+    public static Vector2f getScreenSize() {
         Window window = MinecraftClient.getInstance().getWindow();
-        return new Vector2i(window.getWidth(), window.getHeight());
+        return new Vector2f((float)window.getWidth(), (float)window.getHeight());
     }
     
     public static void setWindow(long window) {
